+++ src/RuleSet/Php56.php
@@ -821,5 +821,5 @@ final class Php56 extends AbstractRuleSet implements ExplicitRuleSet
             'less_and_greater' => true,
         ],
     ];
-    protected int $targetPhpVersion = 70100;
+    protected int $targetPhpVersion = 50600;
 }