+++ test/Unit/RuleSet/Php56Test.php
@@ -826,5 +826,5 @@ final class Php56Test extends ExplicitRuleSetTestCase
             'less_and_greater' => true,
         ],
     ];
-    protected int $targetPhpVersion = 70100;
+    protected int $targetPhpVersion = 50600;
 }