+++ test/Unit/RuleSet/Php70Test.php
@@ -826,5 +826,5 @@ final class Php70Test extends ExplicitRuleSetTestCase
             'less_and_greater' => true,
         ],
     ];
-    protected int $targetPhpVersion = 70100;
+    protected int $targetPhpVersion = 70000;
 }